Updated the led strip driver
Updated the dma controller settings, and changed the ledpixel buffer size to 8 bits instead of 16. this saves 50% size that would be unusable otherwise.

this is done by using dma0 ch0 to initiate the dma transfer request and write a byte from the pixel buffer to an intermediate buffer of size 32,
then when one tranfer is done, a link request to dma0 ch1 is started which initiates a transfer on dma1, which transfers from the intermediate buffer to the CnV register.

this is done because dma doesnt work well with source address increment when the destination address is bigger than the source address.